+++ CHANGELOG.md
@@ -43,3 +43,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Change some validate functions that has `invalidOption` error, to throw error on `invalidOption` instead of returning it.
 - Update regex for UUID
 - Added Justice service error translator library
+- Added translateServiceErrorForAdmin on Justice service error translator library, add some translations